% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-north-arrow.R
\name{annotation_north_arrow}
\alias{annotation_north_arrow}
\alias{GeomNorthArrow}
\title{Spatial-aware north arrow}
\usage{
annotation_north_arrow(
  mapping = NULL,
  data = NULL,
  ...,
  height = unit(1.5, "cm"),
  width = unit(1.5, "cm"),
  pad_x = unit(0.25, "cm"),
  pad_y = unit(0.25, "cm"),
  rotation = NULL,
  style = north_arrow_orienteering
)

GeomNorthArrow
}
\arguments{
\item{mapping, data, ...}{See Aesthetics}

\item{height, width}{Height and width of north arrow}

\item{pad_x, pad_y}{Padding between north arrow and edge of frame}

\item{rotation}{Override the rotation of the north arrow (degrees conterclockwise)}

\item{style}{A grob or callable that produces a grob that will be drawn as the north arrow.
See \link{north_arrow_orienteering} for options.}
}
\value{
A ggplot2 layer
}
\description{
Spatial-aware north arrow
}
\section{Aesthetics}{

The following can be used as parameters or aesthetics. Using them as
aesthetics is useful when facets are used to display multiple panels,
and a different (or missing) scale bar is required in different panels.
Otherwise, just pass them as arguments to \code{annotation_north_arrow()}.
\itemize{
\item which_north: "grid" results in a north arrow always pointing up; "true" always points to the
north pole from whichever corner of the map the north arrow is in.
\item location: Where to put the scale bar ("tl" for top left, etc.)
}
}

\examples{

cities <- data.frame(
  x = c(-63.58595, 116.41214),
  y = c(44.64862, 40.19063),
  city = c("Halifax", "Beijing")
)

ggplot(cities) +
  geom_spatial_point(aes(x, y), crs = 4326) +
  annotation_north_arrow(which_north = "true") +
  coord_sf(crs = 3995)

ggplot(cities) +
  geom_spatial_point(aes(x, y), crs = 4326) +
  annotation_north_arrow(which_north = "grid") +
  coord_sf(crs = 3995)

}
