% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{as_ggspatial_crs}
\alias{as_ggspatial_crs}
\alias{as_ggspatial_crs.default}
\alias{as_ggspatial_crs.sf}
\title{Extract a projection from an object}
\usage{
as_ggspatial_crs(x)

\method{as_ggspatial_crs}{default}(x)

\method{as_ggspatial_crs}{sf}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A CRS object, or NA if one cannot be extracted
}
\description{
Many functions in this package require projection information. This function
allows a CRS to be extracted from various objects, notably Spatial*,
Raster*, and integers (which are assumed to be EPSG codes). This allows
for less verbose syntax when dealing with projections in this package.
}
\examples{
as_ggspatial_crs(4326) # integer
as_ggspatial_crs(longlake_osm) # raster
as_ggspatial_crs(longlake_waterdf) # spatial

library(sf)
as_ggspatial_crs(st_as_sf(longlake_waterdf)) # sf

}
