% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{geom_spatial.Raster}
\alias{geom_spatial.Raster}
\alias{ggraster}
\alias{spatial_fortify.Raster}
\alias{spatial_geom.Raster}
\alias{spatial_default_aes.Raster}
\alias{geom_spatial.sf}
\alias{geom_spatial.sfc}
\alias{spatial_fortify.sf}
\alias{spatial_fortify.sfc}
\alias{spatial_geom.sf}
\alias{spatial_geom.sfc}
\alias{spatial_default_aes.sf}
\alias{spatial_default_aes.sfc}
\alias{spatial_fortify.SpatialPointsDataFrame}
\alias{spatial_fortify.SpatialLinesDataFrame}
\alias{spatial_fortify.SpatialPolygonsDataFrame}
\alias{geom_spatial}
\alias{ggspatial}
\alias{geom_spatial.data.frame}
\alias{geom_spatial.default}
\alias{spatial_fortify}
\alias{spatial_fortify.default}
\alias{stat_project}
\alias{annotation_spraster}
\alias{geom_spraster_rgb}
\alias{geom_osm}
\alias{ggosm}
\title{Deprecated functions}
\usage{
\method{geom_spatial}{Raster}(data, mapping = NULL, show.legend = TRUE,
  inherit.aes = FALSE, position = "identity", crsfrom = NA,
  crsto = NULL, attribute_table = NA, geom = NA, stat = NA, ...)

ggraster(data, mapping = NULL, ...)

\method{spatial_fortify}{Raster}(x, ..., format = c("wide", "long"))

\method{spatial_geom}{Raster}(x)

\method{spatial_default_aes}{Raster}(x)

\method{geom_spatial}{sf}(data, mapping = NULL, show.legend = TRUE,
  inherit.aes = FALSE, position = "identity", crsfrom = NA, crsto = NA,
  attribute_table = NA, geom = NA, stat = NA, ...)

\method{geom_spatial}{sfc}(data, mapping = NULL, show.legend = TRUE,
  inherit.aes = FALSE, position = "identity", crsfrom = NA, crsto = NA,
  attribute_table = NA, geom = NA, stat = NA, ...)

\method{spatial_fortify}{sf}(x, attrs = NA, ...)

\method{spatial_fortify}{sfc}(x, attrs = NA, ...)

\method{spatial_geom}{sf}(x)

\method{spatial_geom}{sfc}(x)

\method{spatial_default_aes}{sf}(x)

\method{spatial_default_aes}{sfc}(x)

\method{spatial_fortify}{SpatialPointsDataFrame}(x, attrs = NA, ...)

\method{spatial_fortify}{SpatialLinesDataFrame}(x, attrs = NA, ...)

\method{spatial_fortify}{SpatialPolygonsDataFrame}(x, attrs = NA, ...)

geom_spatial(data, mapping = NULL, ...)

ggspatial(data, mapping = NULL, ...)

\method{geom_spatial}{data.frame}(data, mapping = NULL, ...)

\method{geom_spatial}{default}(data, mapping = NULL, show.legend = TRUE,
  inherit.aes = FALSE, position = "identity", crsfrom = NA, crsto = NA,
  attribute_table = NA, geom = NA, stat = NA, ...)

spatial_fortify(x, attrs = NULL, ...)

\method{spatial_fortify}{default}(x, attrs = NA, ...)

stat_project(mapping = NULL, data = NULL, crsfrom = NA, crsto = NA,
  position = "identity", show.legend = TRUE, inherit.aes = TRUE,
  geom = "point", ...)

annotation_spraster(raster, interpolate = FALSE, na.value = NA)

geom_spraster_rgb(raster, interpolate = FALSE, na.value = NA)

geom_osm(x = NULL, zoomin = 0, zoom = NULL, type = NULL,
  forcedownload = FALSE, cachedir = NULL, progress = c("text", "none"),
  quiet = TRUE)

ggosm(x = NULL, zoomin = 0, zoom = NULL, type = NULL,
  forcedownload = FALSE, cachedir = NULL, progress = c("text", "none"),
  quiet = TRUE)
}
\arguments{
\item{data, mapping, geom, stat, position, show.legend, inherit.aes}{deprecated}

\item{crsfrom, crsto, attribute_table}{deprecated}

\item{...}{Passed to/from methods}

\item{format, attrs, x, raster, interpolate, na.value}{deprecated}

\item{zoomin, zoom, type, forcedownload, cachedir, progress, quiet}{deprecated}
}
\description{
These functions are deprecated and should not be used.
}
