% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spatial.R
\name{fortify.SpatialPoints}
\alias{fortify.SpatialLines}
\alias{fortify.SpatialPoints}
\alias{fortify.SpatialPointsDataFrame}
\title{Create a data.frame from Spatial* objects}
\usage{
\method{fortify}{SpatialPoints}(model, data = NULL, ...)

\method{fortify}{SpatialPointsDataFrame}(model, data = NULL, ...)

\method{fortify}{SpatialLines}(model, data, ...)
}
\arguments{
\item{model}{The object}

\item{data}{Not used by this method}

\item{...}{Not used by this method}
}
\value{
A data.frame with colummns id, long and lat (mimics
behaviour in fortify.Spatial* from ggplot2.
}
\description{
Methods to turn Spatial* objects into data.frames
}
\details{
Several methods are missing from ggplot2, including \link[ggplot2]{fortify}
implementations for SpatialPoints, SpatialPointsDataFrame, and SpatialLines.
}
\examples{
fortify(longlake_depthdf)

}

