% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmggplot.R
\name{geom_osm}
\alias{geom_osm}
\title{A ggplot geometry for OSM imagery}
\usage{
geom_osm(obj = NULL, zoomin = 0, zoom = NULL, type = "osm",
  forcedownload = FALSE, cachedir = NULL, epsg = 4326, ...)
}
\arguments{
\item{obj}{An object like in \link{osm.raster}: a bounding box or Spatial* object.}

\item{zoomin}{The amount by which to adjust the automatically calculated zoom (or
manually specified if the \code{zoom} parameter is passed). Use +1 to zoom in, or -1 to zoom out.}

\item{zoom}{Manually specify the zoom level (not recommended; adjust \code{zoomin} instead.}

\item{type}{A map type; one of that returned by \link{osm.types}. User defined types are possible
by defining \code{tile.url.TYPENAME <- function(xtile, ytile, zoom){}} and passing TYPENAME
as the \code{type} argument.}

\item{forcedownload}{\code{TRUE} if cached tiles should be re-downloaded. Useful if
some tiles are corrupted.}

\item{cachedir}{The directory in which tiles should be cached. Defaults to \code{getwd()/rosm.cache}.}

\item{epsg}{The epsg code of the projection of the coordinates being plotted by other geoms}

\item{...}{Agruments passed on to \code{geom_raster()}}
}
\value{
A \code{geom_raster} with colour data from the tiles.
}
\description{
An experimental function returning a geom_raster representing the tile data such that it
can be plotted as a \code{ggplot2} layer. Should probably be used with \code{coord_fixed}.
Note that this does not scale the aspect like the \code{sp} package and assumes lat/lon
as default coordinates (specify with )
}
\examples{
\donttest{
library(prettymapr)
library(ggplot2)
ns <- searchbbox("Nova Scotia")

ggplot(data.frame(x=1)) + geom_osm(ns)

ggplot(data.frame(t(ns)), aes(x=x, y=y)) +
  geom_osm(type="osm", zoomin=-1) +
  geom_point() + coord_fixed()
}

}

