% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spraster.R
\name{geom_spatial.Raster}
\alias{geom_spatial.Raster}
\title{Spatial Geometry for Raster* Objects}
\usage{
\method{geom_spatial}{Raster}(data, mapping = NULL, show.legend = TRUE,
  inherit.aes = FALSE, position = "identity", fromepsg = NULL,
  toepsg = NULL, fromprojection = NULL, toprojection = NULL,
  geom = "raster", stat = "identity", ...)
}
\arguments{
\item{data}{A Raster* object}

\item{mapping}{A mapping}

\item{show.legend}{Should the legend be shown for this layer?}

\item{inherit.aes}{Should aesthetics be inherited from the base plot?}

\item{position}{The position to apply (should probably always be 'identity')}

\item{fromepsg}{Override the source projection}

\item{toepsg}{Project cell coordinates to new projection}

\item{fromprojection}{Override the source projection}

\item{toprojection}{Project cell coordinates to new projection}

\item{geom}{The geometry to use. Defaults to raster (obviously), but could also
be another value if used with a different stat (e.g. contour)}

\item{stat}{The stat to apply. Defaults to 'identity', but could be something else
like 'contour' or \link{stat_rgba}.}

\item{...}{Further arguments passed to the stat/geom}
}
\value{
A ggplot2 layer
}
\description{
Similar to the \link{geom_spatial} family for Spatial* objects, this method
plots a spatial raster as a spatial raster. Note that projecting (or un-projecting)
this layer will likely result in odd results (namely that geom_raster will not
function properly or at all).
}
\examples{
ggplot() + geom_spatial(longlake_osm)

}

