% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spggplot.R
\name{geom_spatial}
\alias{geom_spatial}
\title{A ggplot2 geom for Spatial* objects}
\usage{
geom_spatial(data = NULL, mapping = NULL, show.legend = TRUE,
  inherit.aes = TRUE, position = "identity", fromepsg = NULL,
  toepsg = NULL, fromprojection = NULL, toprojection = NULL, ...)
}
\arguments{
\item{data}{A \code{Spatial*} object or \code{data.frame}.}

\item{mapping}{A mapping as created by \code{aes()} or \code{aes_string()}}

\item{show.legend}{Logical describing the legend visibility.}

\item{inherit.aes}{Logical describing if aesthetics are inherited}

\item{position}{Passed on to geom_*}

\item{fromepsg}{The epsg code of the projection of the dataset (defaults to lat/lon or 4326 if
not specified or the \code{data} does not specify a CRS)}

\item{toepsg}{The projection of the final plot (defaults to \code{fromepsg} if not specified
or 3857 if \code{fromepsg} is not specified)}

\item{fromprojection}{Long form of \code{fromepsg}, a CRS object created by \code{sp::CRS()}}

\item{toprojection}{Long form of \code{toepsg}, a CRS object created by \code{sp::CRS()}}

\item{...}{Agruments passed on to the \code{geom_*} (e.g. \code{lwd}, \code{fill}, etc.)}
}
\description{
A function returning a geom_* object based on the Spatial* input. Also will
happily project a regular \code{data.frame} provided x and y aesthetics are
specified. The result is a \code{geom_*} for use with ggplot2, with aesthetics
and other argumets passed on to that geom.
}
\examples{
\donttest{
library(prettymapr)
ns <- searchbbox("Nova Scotia")
cities <- geocode(c("Wolfville, NS", "Windsor, NS", "Halifax, NS"))
ggplot(cities, aes(x=lon, y=lat)) + geom_spatial(toepsg=26920) + coord_fixed()
# default projection is Spherical Mercator (EPSG:3857)
ggplot(cities, aes(x=lon, y=lat)) + geom_spatial() + coord_fixed()

# plot a number of spatial objects
data(longlake_marshdf, longlake_roadsdf, longlake_waterdf, longlake_depthdf,
     longlake_streamsdf, longlake_buildingsdf)
ggplot() +
  geom_spatial(longlake_waterdf, fill="lightblue") +
  geom_spatial(longlake_marshdf, fill="grey", alpha=0.5) +
  geom_spatial(longlake_streamsdf, col="lightblue") +
  geom_spatial(longlake_roadsdf, col="black") +
  geom_spatial(longlake_buildingsdf, pch=1, col="brown", size=0.25) +
  geom_spatial(longlake_depthdf, aes(col=DEPTH.M)) +
  facet_wrap(~NOTES)+
  coord_fixed()
}

}

