% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{geom_spatial.sf}
\alias{geom_spatial.sf}
\alias{geom_spatial.sfc}
\title{A ggplot2 geom for sf* objects}
\usage{
\method{geom_spatial}{sf}(data, mapping = NULL, show.legend = TRUE,
  inherit.aes = FALSE, position = "identity", crsfrom = NA, crsto = NA,
  attribute_table = NA, geom = NA, stat = NA, ...)

\method{geom_spatial}{sfc}(data, mapping = NULL, show.legend = TRUE,
  inherit.aes = FALSE, position = "identity", crsfrom = NA, crsto = NA,
  attribute_table = NA, geom = NA, stat = NA, ...)
}
\arguments{
\item{data}{A \code{Spatial*} object or \code{data.frame}.}

\item{mapping}{A mapping as created by \code{aes()} or \code{aes_string()}}

\item{show.legend}{Logical describing the legend visibility.}

\item{inherit.aes}{Logical describing if aesthetics are inherited}

\item{position}{Passed on to the layer}

\item{crsfrom}{An object that can be coerced to a CRS using \link{as.CRS}; defaults
to the CRS of the data or lat/lon if that does not exist}

\item{crsto}{An object that can be coerced to a CRS using \link{as.CRS}; defaults to
lat/lon so that the plot can be projected using coord_map()}

\item{attribute_table}{For SpatialPoints, SpatialLines, and SpatialPolygons, an attribute
table that matches the input object.}

\item{geom}{The geometry to use for the object (NA to guess: see \link{spatial_geom})}

\item{stat}{The statistic to apply (NA to guess, is probably "identity": see \link{spatial_stat})}

\item{...}{Agruments passed on to the \code{geom_*} (e.g. \code{lwd}, \code{fill}, etc.)}
}
\value{
A ggplot2 Geom object
}
\description{
This implementation of \link{geom_spatial} for the sf package currently converts
objects to a Spatial* object, although this will likely change in future versions.
}
\examples{
library(sf)
llsf <- st_as_sf(longlake_waterdf)
ggspatial(llsf)
llsfc <- st_as_sfc(longlake_waterdf)
ggspatial(llsfc)

}
