% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraster.R
\name{geom_spraster}
\alias{geom_spraster}
\title{Get a spatial raster as a geom_raster for ggplot2}
\usage{
geom_spraster(rastobj, scale = 1, rgb = NULL, band = NULL,
  aesthetic = "fill", show.legend = FALSE, ...)
}
\arguments{
\item{rastobj}{A Raster, RasterStack, or RasterBrick}

\item{scale}{The scaling to apply to the raster before applying rgb conversion or
aesthetics. May need to use 1/255 for alpha.}

\item{rgb}{The bands for which rgb should be computed (1:3 for bands >= 3). Pass
FALSE if rgb should not be converted.}

\item{band}{The band to be used for aesthetic should more than one be available.}

\item{aesthetic}{Should probably always be "fill" but could be "alpha"}

\item{...}{Agruments passed on to geom_raster.}
}
\value{
A geom_raster().
}
\description{
Note that rgb and alpha is not yet supported.
}
\examples{
data(longlake_osm)
ggplot() + geom_spraster(longlake_osm)
ggplot() + geom_spraster(longlake_osm, rgb=c(3,2,1))
ggplot() + geom_spraster(longlake_osm, band=2)
ggplot() + geom_spraster(longlake_osm, band=2, aesthetic="alpha")

}

