% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggraticule.R
\name{ggraticule}
\alias{ggraticule}
\title{Prepare graticule object for ggplot of spatial objects}
\usage{
ggraticule(x, ext = extent(x))
}
\arguments{
\item{x}{Projected Spatial object}

\item{ext}{extent Extent object for the plot (default is to use the extent of x)}
}
\description{
None
}
\author{
Loic Dutrieux
}

