% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggRGB.R
\name{ggRGB}
\alias{ggRGB}
\title{RGB spatial plot with ggplot}
\usage{
ggRGB(x, r, g, b)
}
\arguments{
\item{x}{RasterStackBrick object}

\item{r}{Numeric Band number to be plotted in red}

\item{g}{Numeric Band number to be plotted in green}

\item{b}{Numeric Band number to be plotted in blue}
}
\description{
Produces an object to feed into ggplot geom_raster or geom_tile to produce a rgb plot
}
\examples{
library(ggplot2)
library(raster)

x <- brick(system.file("external/rlogo.grd", package="raster"))

ggrgb <- ggRGB(x, 3,2,1)

ggplot() +
    geom_raster(data = ggrgb, aes(x=x, y=y, fill=values)) +
    scale_fill_identity() +
    coord_equal() +
    scale_x_continuous(name = 'Lat', expand=c(0, 0)) +
    scale_y_continuous(name = 'Long', expand=c(0, 0))
}
\author{
Loic Dutrieux (using some code snippets by Josh Gray and Robert Hijmans)
}

