% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-spatial-raster.R
\name{layer_spatial.Raster}
\alias{layer_spatial.Raster}
\alias{annotation_spatial.Raster}
\alias{StatSpatialRaster}
\alias{StatSpatialRasterAnnotation}
\alias{StatSpatialRasterDf}
\alias{GeomSpatialRaster}
\title{Spatial ggplot2 layer for raster objects}
\usage{
\method{layer_spatial}{Raster}(
  data,
  mapping = NULL,
  interpolate = NULL,
  is_annotation = FALSE,
  lazy = FALSE,
  dpi = 150,
  ...
)

\method{annotation_spatial}{Raster}(data, mapping = NULL, interpolate = NULL, ...)

StatSpatialRaster

StatSpatialRasterAnnotation

StatSpatialRasterDf

GeomSpatialRaster
}
\arguments{
\item{data}{A Raster object}

\item{mapping}{Currently, only RGB or RGBA rasters are supported. In the future, one may be able to
map specific bands to the fill and alpha aesthetics.}

\item{interpolate}{Interpolate resampling for rendered raster image}

\item{is_annotation}{Lets raster exist without modifying scales}

\item{lazy}{Delay projection and resample of raster until the plot is being rendered}

\item{dpi}{if lazy = TRUE, the dpi to which the raster should be resampled}

\item{...}{Passed to other methods}
}
\value{
A ggplot2 layer
}
\description{
This is intended for use with RGB(A) rasters (e.g., georeferenced imagery or photos). To work with
bands as if they were columns, use \code{\link[=df_spatial]{df_spatial()}} and \code{\link[ggplot2:geom_raster]{ggplot2::geom_raster()}}.
}
\examples{
\donttest{
library(ggplot2)
load_longlake_data(which = c("longlake_osm", "longlake_depth_raster"))
ggplot() + layer_spatial(longlake_osm)
ggplot() + layer_spatial(longlake_depth_raster) + scale_fill_continuous(na.value = NA)
}

}
