% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlake-data.R
\name{load_longlake_data}
\alias{load_longlake_data}
\title{Load longlake test data}
\source{
The Nova Scotia Topographic Database
(\url{https://geonova.novascotia.ca/}) and
Open Street Map (\url{http://www.openstreetmap.org}).
}
\usage{
load_longlake_data(env = parent.frame())
}
\arguments{
\item{env}{The environment in which to assign the objects}
}
\description{
Load longlake test data
}
\examples{
load_longlake_data()

}
