% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longlake_buildingsdf}
\alias{longlake_buildingsdf}
\alias{longlake_depthdf}
\alias{longlake_marshdf}
\alias{longlake_osm}
\alias{longlake_roadsdf}
\alias{longlake_streamsdf}
\alias{longlake_waterdf}
\title{Long Lake Spatial* objects}
\format{An object of class \code{SpatialPointsDataFrame} with 14 rows and 16 columns.}
\source{
The Nova Scotia Topographic Database
(\url{http://www.mapsnovascotia.com/category.aspx?ic=24}) and
Open Street Map (\url{http://www.openstreetmap.org}).
}
\usage{
longlake_buildingsdf

longlake_buildingsdf

longlake_depthdf

longlake_marshdf

longlake_osm

longlake_roadsdf

longlake_streamsdf

longlake_waterdf
}
\description{
A group of Spatial* objects used to test \link{geom_spatial}.
}
\keyword{datasets}

