% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-north-arrow.R
\name{north_arrow_orienteering}
\alias{north_arrow_orienteering}
\alias{north_arrow_fancy_orienteering}
\alias{north_arrow_minimal}
\alias{north_arrow_nautical}
\title{North arrow styles}
\usage{
north_arrow_orienteering(
  line_width = 1,
  line_col = "black",
  fill = c("white", "black"),
  text_col = "black",
  text_family = "",
  text_face = NULL,
  text_size = 10,
  text_angle = 0
)

north_arrow_fancy_orienteering(
  line_width = 1,
  line_col = "black",
  fill = c("white", "black"),
  text_col = "black",
  text_family = "",
  text_face = NULL,
  text_size = 10,
  text_angle = 0
)

north_arrow_minimal(
  line_width = 1,
  line_col = "black",
  fill = "black",
  text_col = "black",
  text_family = "",
  text_face = NULL,
  text_size = 10
)

north_arrow_nautical(
  line_width = 1,
  line_col = "black",
  fill = c("black", "white"),
  text_size = 10,
  text_face = NULL,
  text_family = "",
  text_col = "black",
  text_angle = 0
)
}
\arguments{
\item{line_width, line_col, fill}{Parameters customizing the appearance of the north arrow}

\item{text_col, text_family, text_face, text_size, text_angle}{Parameters customizing the
text of the north arrow}
}
\value{
A Grob with npc coordinates (more or less) 0 to 1
}
\description{
North arrow styles
}
\examples{
grid::grid.newpage()
grid::grid.draw(north_arrow_orienteering())

grid::grid.newpage()
grid::grid.draw(north_arrow_fancy_orienteering())

grid::grid.newpage()
grid::grid.draw(north_arrow_minimal())

grid::grid.newpage()
grid::grid.draw(north_arrow_nautical())

}
