% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmggplot.R
\name{osm.ggplot}
\alias{osm.ggplot}
\title{A ggplot interface for OSM imagery}
\usage{
osm.ggplot(x, zoomin = 0, zoom = NULL, type = "osm",
  forcedownload = FALSE, cachedir = NULL, epsg = 3857, ...)
}
\arguments{
\item{x}{A bounding box as generated by \code{sp::bbox()} or \code{prettymapr::searchbbox()}.
Must be in lon/lat (epsg:4326)! Alternatively, pass a \code{Spatial*} object to use
the bounding box of that}

\item{zoomin}{The amount by which to adjust the automatically calculated zoom (or
manually specified if the \code{zoom} parameter is passed). Use +1 to zoom in, or -1 to zoom out.}

\item{zoom}{Manually specify the zoom level (not recommended; adjust \code{zoomin} instead.}

\item{type}{A map type; one of that returned by \link{osm.types}. User defined types are possible
by defining \code{tile.url.TYPENAME <- function(xtile, ytile, zoom){}} and passing TYPENAME
as the \code{type} argument.}

\item{forcedownload}{\code{TRUE} if cached tiles should be re-downloaded. Useful if
some tiles are corrupted.}

\item{cachedir}{The directory in which tiles should be cached. Defaults to \code{getwd()/rosm.cache}.}

\item{epsg}{The epsg code of the projection of the coordinates being plotted by other geoms.
Note that if this is not a cylindrical projection, this will not be entirely accurate.}

\item{...}{Arguments passed on to \code{raster::writeRaster()} if \code{filename} is specified.}
}
\value{
A \code{geom_raster} with colour data from the tiles.
}
\description{
An experimental function returning a \code{ggplot} object more or less equivalent to the
\link{osm.plot} function. This is considerably less efficient than using \link{osm.plot}
but may be desirable if 'ggplot' is the plotting framework of choice.
}
\examples{
\donttest{
library(prettymapr)
ns <- searchbbox("Nova Scotia")
osm.ggplot(ns)

}
}
\seealso{
geom_osm, rosm::osm.raster
}

