% Generated by roxygen2 (4.0.2): do not edit by hand
\name{raster2ggdf}
\alias{raster2ggdf}
\title{Prepare a raster object to plot using geom_raster or geom_tile}
\usage{
raster2ggdf(x)
}
\arguments{
\item{x}{raster object}
}
\description{
Prepare a raster object to plot using geom_raster or geom_tile
}
\examples{
library(raster)
library(ggplot2)

r <- raster(nrow = 10, ncol = 20)
r[] <- rnorm(ncell(r))

ggr <- fortify(r)

ggplot() +
    geom_raster(data = ggr, aes(x,y,fill = values)) +
    coord_equal() +
    scale_x_continuous(name = 'Lat', expand=c(0, 0)) +
    scale_y_continuous(name = 'Long', expand=c(0, 0))

# Also works for rasterStackBricks

b <- brick(nrow = 10, ncol = 20, nl = 3)
b[] <- rnorm(ncell(b) * nlayers(b))
names(b) <- c('a', 'b', 'c')
ggb <- fortify(b)

ggplot() +
    geom_raster(data = ggb, aes(x,y,fill = values.a)) +
    coord_equal() +
    scale_x_continuous(name = 'Lat', expand=c(0, 0)) +
    scale_y_continuous(name = 'Long', expand=c(0, 0))

# For facet grid/wrap, you can use tidyr to reshape the dataframe produced by the function
library(tidyr)

ggb2 <- gather(ggb, layer, values, -x, -y)

ggplot() +
    geom_raster(data = ggb2, aes(x,y,fill = values)) +
    coord_equal() +
    scale_x_continuous(name = 'Lat', expand=c(0, 0)) +
    scale_y_continuous(name = 'Long', expand=c(0, 0))+
    facet_wrap(~ layer)
}
\author{
Loic Dutrieux
}

