% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspatial-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplot}
\alias{aes}
\alias{coord_sf}
\alias{geom_sf}
\alias{waiver}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:aes]{aes()}}, \code{\link[ggplot2:coord_sf]{coord_sf()}}, \code{\link[ggplot2:geom_sf]{geom_sf()}}, \code{\link[ggplot2:ggplot]{ggplot()}}, \code{\link[ggplot2:waiver]{waiver()}}}
}}

