% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{spatial_fortify.Raster}
\alias{spatial_fortify.Raster}
\alias{spatial_fortify.sf}
\alias{spatial_fortify.sfc}
\alias{spatial_fortify.SpatialPointsDataFrame}
\alias{spatial_fortify.SpatialLinesDataFrame}
\alias{spatial_fortify.SpatialPolygonsDataFrame}
\alias{spatial_fortify}
\alias{spatial_fortify.default}
\title{Create a data frame from a Spatial object with attributes}
\usage{
\method{spatial_fortify}{Raster}(x, ...)

\method{spatial_fortify}{sf}(x, attrs = NA, ...)

\method{spatial_fortify}{sfc}(x, attrs = NA, ...)

\method{spatial_fortify}{SpatialPointsDataFrame}(x, attrs = NA, ...)

\method{spatial_fortify}{SpatialLinesDataFrame}(x, attrs = NA, ...)

\method{spatial_fortify}{SpatialPolygonsDataFrame}(x, attrs = NA, ...)

spatial_fortify(x, attrs = NULL, ...)

\method{spatial_fortify}{default}(x, attrs = NA, ...)
}
\arguments{
\item{x}{A spatial object}

\item{...}{Additional arguments passed to \link[ggplot2]{fortify}}

\item{attrs}{An optional attribute table, NA for the default, or NULL for none.}
}
\value{
A data.frame with (at least) columns ".lat" and ".long".
}
\description{
The \link[ggplot2]{fortify} function returns a data frame with geometry information;
\code{spatial_fortify()} returns a data frame with the results of fortify (left) joined
with the attribute table. If there is no attribute table, the results of fortify are
returned with the column names preceeded by a \code{.}.
}
\examples{
head(spatial_fortify(longlake_depthdf))

}
