% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spraster.R
\name{stat_rgba}
\alias{stat_rgba}
\title{Statistic to map rgb values}
\usage{
stat_rgba(mapping = NULL, data = NULL, geom = "point",
  position = "identity", ...)
}
\arguments{
\item{mapping}{A mapping created with \link[ggplot2]{aes}}

\item{data}{A data.frame}

\item{geom}{The geometry to use}

\item{position}{The position adjustment to use}

\item{...}{Aesthetics or parameters to set}
}
\value{
A ggplot2 layer object
}
\description{
Statistic to map rgb values
}

