% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-project.R
\docType{data}
\name{StatProject}
\alias{StatProject}
\title{Statisitc to project coordinates}
\format{An object of class \code{StatProject} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatProject
}
\description{
Projects coordinates using rgdal/sp. For use with \link{geom_spatial}. Takes params
\code{crsfrom} and \code{crsto}, both generated by \code{sp::CRS()} or \code{NULL}.
If \code{NULL}, \code{crsto} is assumed to be EPSG:3857 (Spherical Mercator) and
\code{crsto} is assumed to be EPSG:4326 (Lat/Lon).
}
\keyword{datasets}

