% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spatial.R
\name{xy_transform}
\alias{xy_transform}
\title{Coordinate transform}
\usage{
xy_transform(x, y, from = 4326, to = 4326, na.rm = FALSE)
}
\arguments{
\item{x}{The x coordinate}

\item{y}{The y coordinate}

\item{from}{From CRS}

\item{to}{To CRS}

\item{na.rm}{Warn for non-finite cases?}
}
\value{
A data.frame with x and y components.
}
\description{
Coordinate transform, propotating non-finite cases.
}
\examples{
xy_transform(c(1, 2, 3), c(1, 2, 3), to = 3857)
xy_transform(c(1, 2, 3), c(NA, NA, NA), to = 3857)
xy_transform(c(1, 2, 3), c(NA, 2, 3), to = 3857)
xy_transform(c(1, 2, 3), c(1, 2, NA), to = 3857)

}
