% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-map-tile.R
\name{annotation_map_tile}
\alias{annotation_map_tile}
\alias{GeomMapTile}
\title{Add background OSM tiles}
\usage{
annotation_map_tile(
  type = "osm",
  zoom = NULL,
  zoomin = -2,
  forcedownload = FALSE,
  cachedir = NULL,
  progress = c("text", "none"),
  quiet = TRUE,
  interpolate = TRUE,
  data = NULL,
  mapping = NULL,
  alpha = 1,
  api_key = NULL
)

GeomMapTile
}
\arguments{
\item{type}{The map type (one of that returned by \link[rosm:osm.types]{rosm::osm.types})}

\item{zoom}{The zoom level (overrides zoomin)}

\item{zoomin}{Delta on default zoom. The default value is designed
to download fewer tiles than you probably want. Use \code{-1} or \code{0} to
increase the resolution.}

\item{forcedownload}{Re-download cached tiles?}

\item{cachedir}{Specify cache directory}

\item{progress}{Use \code{progress = "none"} to suppress progress and zoom output}

\item{quiet}{Use \code{quiet = FALSE} to see which URLs are downloaded}

\item{interpolate}{Passed to \code{\link[grid:rasterGrob]{grid::rasterGrob()}}}

\item{data, mapping}{Specify data and mapping to use this geom with facets}

\item{alpha}{Use to make this layer semi-transparent}

\item{api_key}{API key used for CARTO map types. If \code{NULL}, rosm uses the
\code{CARTO_API_KEY} environment variable.}
}
\value{
A ggplot2 layer
}
\description{
Uses \code{\link[rosm:osm.image]{rosm::osm.image()}} to add background tiles. If you are publishing
a map using these tiles, make sure to use the proper attribution
(e.g., "Copyright OpenStreetMap contributors" when using an
OpenStreetMap-based tile set).
}
\examples{
\donttest{
library(ggplot2)
load_longlake_data(which = "longlake_waterdf")

ggplot() +
  annotation_map_tile(zoom = 13, cachedir = system.file("rosm.cache", package = "ggspatial")) +
  geom_sf(data = longlake_waterdf, fill = NA, col = "grey50")
}

}
