% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spatial-xline.R
\docType{data}
\name{annotation_spatial_hline}
\alias{annotation_spatial_hline}
\alias{annotation_spatial_vline}
\alias{GeomSpatialXline}
\title{Projected horizontal and vertical lines}
\format{An object of class \code{GeomSpatialXline} (inherits from \code{GeomHline}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
annotation_spatial_hline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  ...,
  intercept = waiver(),
  limits = NULL,
  detail = 100,
  na.rm = FALSE,
  show.legend = NA
)

annotation_spatial_vline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  ...,
  intercept = waiver(),
  limits = NULL,
  detail = 100,
  na.rm = FALSE,
  show.legend = NA
)

GeomSpatialXline
}
\arguments{
\item{intercept}{The x or y value that should be constant in the given
`crs`. Can also be passed as an aesthetic through `data` and `mapping`.}

\item{limits}{Use `NULL` to guess the minimum and maximum x or y value in
the non-constant dimension, or specify a vector of length 2 to specify
manually.}

\item{detail}{The number of points that should be used when converting the
line into segments.}
}
\description{
Projected horizontal and vertical lines
}
\keyword{datasets}
