% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spatial.R
\name{geom_spatial}
\alias{geom_spatial}
\alias{geom_spatial.SpatialLines}
\alias{geom_spatial.SpatialLinesDataFrame}
\alias{geom_spatial.SpatialPoints}
\alias{geom_spatial.SpatialPointsDataFrame}
\alias{geom_spatial.SpatialPolygons}
\alias{geom_spatial.SpatialPolygonsDataFrame}
\alias{geom_spatial.default}
\alias{ggspatial}
\title{A ggplot2 geom for Spatial* objects}
\usage{
geom_spatial(data, ...)

ggspatial(data, mapping = NULL, ...)

\method{geom_spatial}{default}(data, mapping = NULL, show.legend = TRUE,
  inherit.aes = NULL, position = "identity", fromepsg = NULL,
  toepsg = NULL, fromprojection = NULL, toprojection = NULL,
  geom = "point", ...)

\method{geom_spatial}{SpatialPoints}(data, mapping = NULL,
  show.legend = TRUE, inherit.aes = NULL, position = "identity",
  fromepsg = NULL, toepsg = NULL, fromprojection = NULL,
  toprojection = NULL, attribute_table = NULL, ...)

\method{geom_spatial}{SpatialPointsDataFrame}(data, mapping = NULL,
  show.legend = TRUE, inherit.aes = NULL, position = "identity",
  fromepsg = NULL, toepsg = NULL, fromprojection = NULL,
  toprojection = NULL, ...)

\method{geom_spatial}{SpatialLines}(data, mapping = NULL,
  show.legend = TRUE, inherit.aes = NULL, position = "identity",
  fromepsg = NULL, toepsg = NULL, fromprojection = NULL,
  toprojection = NULL, attribute_table = NULL, ...)

\method{geom_spatial}{SpatialLinesDataFrame}(data, mapping = NULL,
  show.legend = TRUE, inherit.aes = NULL, position = "identity",
  fromepsg = NULL, toepsg = NULL, fromprojection = NULL,
  toprojection = NULL, ...)

\method{geom_spatial}{SpatialPolygons}(data, mapping = NULL,
  show.legend = TRUE, inherit.aes = NULL, position = "identity",
  fromepsg = NULL, toepsg = NULL, fromprojection = NULL,
  toprojection = NULL, rule = "winding", attribute_table = NULL, ...)

\method{geom_spatial}{SpatialPolygonsDataFrame}(data, mapping = NULL,
  show.legend = TRUE, inherit.aes = NULL, position = "identity",
  fromepsg = NULL, toepsg = NULL, fromprojection = NULL,
  toprojection = NULL, rule = "winding", ...)
}
\arguments{
\item{data}{A \code{Spatial*} object or \code{data.frame}.}

\item{...}{Agruments passed on to the \code{geom_*} (e.g. \code{lwd}, \code{fill}, etc.)}

\item{mapping}{A mapping as created by \code{aes()} or \code{aes_string()}}

\item{show.legend}{Logical describing the legend visibility.}

\item{inherit.aes}{Logical describing if aesthetics are inherited}

\item{position}{Passed on to geom_*}

\item{fromepsg}{The epsg code of the projection of the dataset (defaults to lat/lon or 4326 if
not specified or the \code{data} does not specify a CRS)}

\item{toepsg}{The projection of the final plot (defaults to \code{fromepsg} if not specified
or 3857 if \code{fromepsg} is not specified)}

\item{fromprojection}{Long form of \code{fromepsg}, a CRS object created by \code{sp::CRS()}}

\item{toprojection}{Long form of \code{toepsg}, a CRS object created by \code{sp::CRS()}}

\item{geom}{For data frames, the geometry to use}

\item{attribute_table}{For SpatialPoints, SpatialLines, and SpatialPolygons, an attribute
table that matches the input object.}

\item{rule}{One of 'evenodd' or 'winding', if the Spatial object is a polygon layer.}
}
\value{
A ggplot2 'layer' object.
}
\description{
A function returning a geom_* object based on the Spatial* input. Also will
happily project a regular \code{data.frame} provided x and y aesthetics are
specified. The result is a \code{geom_*} for use with ggplot2, with aesthetics
and other argumets passed on to that geom.
}
\examples{
\donttest{
library(prettymapr)
ns <- searchbbox("Nova Scotia")
cities <- geocode(c("Wolfville, NS", "Windsor, NS", "Halifax, NS"))
ggplot(cities, aes(x=lon, y=lat)) + geom_spatial(toepsg=26920)
# default projection is Spherical Mercator (EPSG:3857)
ggplot(cities, aes(x=lon, y=lat)) + geom_spatial() + coord_map()
}

# plot a number of spatial objects
data(longlake_marshdf, longlake_roadsdf, longlake_waterdf, longlake_depthdf,
     longlake_streamsdf, longlake_buildingsdf)
ggplot() +
  geom_spatial(longlake_waterdf, fill="lightblue") +
  geom_spatial(longlake_marshdf, fill="grey", alpha=0.5) +
  geom_spatial(longlake_streamsdf, col="lightblue") +
  geom_spatial(longlake_roadsdf, col="black") +
  geom_spatial(longlake_buildingsdf, pch=1, col="brown", size=0.25) +
  geom_spatial(longlake_depthdf, aes(col=DEPTH.M)) +
  facet_wrap(~NOTES)+
  coord_map()


library(maptools)
data(wrld_simpl)
ggplot() + geom_spatial(wrld_simpl)

}

