\name{ggSpatial-package}
\alias{ggSpatial-package}
\alias{ggSpatial}
\docType{package}
\title{
Prepare spatial objects for plotting with ggplot2
}
\description{
Mostly a bunch of wrapper to prepare input dataframes to feed into ggplot2. Makes plotting of spatial classes with ggplot2 easier and prettier.
}
\author{
Loic Dutrieux
}
