% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspatial-package.R
\docType{package}
\name{ggspatial-package}
\alias{ggspatial}
\alias{ggspatial-package}
\title{ggspatial: Spatial Data Framework for ggplot2}
\description{
Spatial data plus the power of the ggplot2 framework means easier mapping when input 
  data are already in the form of spatial objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/ggspatial/}
  \item \url{https://github.com/paleolimbot/ggspatial}
  \item Report bugs at \url{https://github.com/paleolimbot/ggspatial/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Other contributors:
\itemize{
  \item Brent Thorne (\href{https://orcid.org/0000-0002-1099-3857}{ORCID}) [contributor]
}

}
\keyword{internal}
