% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-spatial-bbox.R
\name{layer_spatial.bbox}
\alias{layer_spatial.bbox}
\alias{annotation_spatial.bbox}
\title{Add a bounding box to a map}
\usage{
\method{layer_spatial}{bbox}(data, mapping = aes(), ...)

\method{annotation_spatial}{bbox}(data, mapping = aes(), ...)
}
\arguments{
\item{data}{A bounding box generated by [sf::st_bbox()]}

\item{mapping}{A mapping, created using \link[ggplot2]{aes}.}

\item{...}{Passed to \link[ggplot2]{geom_sf}}
}
\description{
To include a bounding box without drawing it, use [shadow_spatial()] on the
original object.
}
\examples{
library(ggplot2)
load_longlake_data()
ggplot() +
  layer_spatial(sf::st_bbox(longlake_waterdf)) +
  layer_spatial(longlake_depthdf)

# use shadow_spatial() to include the geographic area of an object
# without drawing it
ggplot() +
  shadow_spatial(longlake_waterdf) +
  layer_spatial(longlake_depthdf)

}
