% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-spatial-raster.R
\name{layer_spatial.Raster}
\alias{layer_spatial.Raster}
\title{Spatial ggplot2 layer for raster objects}
\usage{
\method{layer_spatial}{Raster}(data, mapping = waiver(), interpolate = TRUE,
  ...)
}
\arguments{
\item{data}{A Raster object}

\item{mapping}{A mapping, using band names or band1, band2, etc. to refer to band names}

\item{interpolate}{Interpolate resampling for rendered raster image}

\item{...}{Passed to other methods}
}
\value{
A ggplot2 layer
}
\description{
Spatial ggplot2 layer for raster objects
}
