% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-spatial-raster.R
\name{layer_spatial.Raster}
\alias{layer_spatial.Raster}
\alias{annotation_spatial.Raster}
\title{Spatial ggplot2 layer for raster objects}
\usage{
\method{layer_spatial}{Raster}(data, mapping = NULL, interpolate = TRUE,
  is_annotation = FALSE, ...)

\method{annotation_spatial}{Raster}(data, mapping = NULL,
  interpolate = TRUE, ...)
}
\arguments{
\item{data}{A Raster object}

\item{mapping}{Currently, only RGB or RGBA rasters are supported. In the future, one may be able to
map specific bands to the fill and alpha aesthetics.}

\item{interpolate}{Interpolate resampling for rendered raster image}

\item{is_annotation}{Lets raster exist without modifying scales}

\item{...}{Passed to other methods}
}
\value{
A ggplot2 layer
}
\description{
This is intended for use with RGB(A) rasters (e.g., georeferenced imagery or photos). To work with
bands as if they were columns, use \link{df_spatial} and \link{geom_raster}.
}
