% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-spatial-raster.R
\docType{data}
\name{layer_spatial.Raster}
\alias{layer_spatial.Raster}
\alias{annotation_spatial.Raster}
\alias{StatSpatialRaster}
\alias{StatSpatialRasterAnnotation}
\alias{StatSpatialRasterDf}
\alias{GeomSpatialRaster}
\title{Spatial ggplot2 layer for raster objects}
\format{
An object of class \code{StatSpatialRaster} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatSpatialRaster} (inherits from \code{StatSpatialRaster}, \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{StatSpatialRasterDf} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomSpatialRaster} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
\method{layer_spatial}{Raster}(
  data,
  mapping = NULL,
  interpolate = NULL,
  is_annotation = FALSE,
  lazy = FALSE,
  dpi = 150,
  ...
)

\method{annotation_spatial}{Raster}(data, mapping = NULL, interpolate = NULL, ...)

StatSpatialRaster

StatSpatialRasterAnnotation

StatSpatialRasterDf

GeomSpatialRaster
}
\arguments{
\item{data}{A Raster object}

\item{mapping}{Currently, only RGB or RGBA rasters are supported. In the future, one may be able to
map specific bands to the fill and alpha aesthetics.}

\item{interpolate}{Interpolate resampling for rendered raster image}

\item{is_annotation}{Lets raster exist without modifying scales}

\item{lazy}{Delay projection and resample of raster until the plot is being rendered}

\item{dpi}{if lazy = TRUE, the dpi to which the raster should be resampled}

\item{...}{Passed to other methods}
}
\value{
A ggplot2 layer
}
\description{
This is intended for use with RGB(A) rasters (e.g., georeferenced imagery or photos). To work with
bands as if they were columns, use \link{df_spatial} and \link{geom_raster}.
}
\examples{
library(ggplot2)
load_longlake_data()
ggplot() + layer_spatial(longlake_osm)
ggplot() + layer_spatial(longlake_depth_raster) + scale_fill_continuous(na.value = NA)

}
\keyword{datasets}
