% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlake-data.R
\name{load_longlake_data}
\alias{load_longlake_data}
\title{Load longlake test data}
\source{
The Nova Scotia Topographic Database
(\url{https://geonova.novascotia.ca/}) and
Open Street Map (\url{https://www.openstreetmap.org/}).
}
\usage{
load_longlake_data(
  env = parent.frame(),
  vector_format = c("sf", "sp"),
  raster_format = c("raster", "stars", "stars_proxy"),
  which = NULL
)
}
\arguments{
\item{env}{The environment in which to assign the objects}

\item{vector_format, raster_format}{The format in which objects should be loaded}

\item{which}{An optional subset of objects to be loaded}
}
\description{
Load longlake test data
}
\examples{
load_longlake_data()

}
