% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spatial.R
\name{stat_spatial_identity}
\alias{stat_spatial_identity}
\alias{geom_spatial_point}
\alias{geom_spatial_path}
\alias{geom_spatial_polygon}
\title{Spatial-aware ggplot2 layers}
\usage{
stat_spatial_identity(mapping = NULL, data = NULL, crs = NULL,
  geom = "point", position = "identity", ..., show.legend = NA,
  inherit.aes = TRUE)

geom_spatial_point(mapping = NULL, data = NULL, crs = NULL, ...)

geom_spatial_path(mapping = NULL, data = NULL, crs = NULL, ...)

geom_spatial_polygon(mapping = NULL, data = NULL, crs = NULL, ...)
}
\arguments{
\item{mapping}{An aesthetic mapping created with \link[ggplot2]{aes}.}

\item{data}{A data frame or other object, coerced to a data.frame by \link[ggplot2]{fortify}.}

\item{crs}{The crs of the x and y aesthetics, or NULL to use default lon/lat
crs.}

\item{geom}{The geometry to use.}

\item{position}{The position to use.}

\item{...}{Passed to the base ggplot2 functions \link[ggplot2]{geom_point},
\link[ggplot2]{geom_path}, and \link[ggplot2]{geom_polygon}, respectively.}

\item{show.legend, inherit.aes}{See \link[ggplot2]{layer}.}
}
\value{
A ggplot2 layer.
}
\description{
These layers are much like their counterparts, \link[ggplot2]{stat_identity},
\link[ggplot2]{geom_point}, \link[ggplot2]{geom_path},
and \link[ggplot2]{geom_polygon}, except they have a \code{crs} argument that
ensures they are projected when using \link[ggplot2]{coord_sf}. Stats are applied to the x and y coordinates
that have been transformed.
}
\examples{
cities <- data.frame(
  x = c(-63.58595, 116.41214, 0),
  y = c(44.64862, 40.19063, 89.9),
  city = c("Halifax", "Beijing", "North Pole")
)

library(ggrepel)
ggplot(cities, aes(x, y)) +
  geom_spatial_point(crs = 4326) +
  stat_spatial_identity(aes(label = city), geom = "label_repel") +
  coord_sf(crs = 3857)

}
